# Try to remove a few deepcopy