--- a/lib/spack/spack/test/cmd/config.py
+++ b/lib/spack/spack/test/cmd/config.py
@@ -640,4 +640,4 @@ spack:
         config("update", "-y", "config")
 
     with ev.Environment(str(tmpdir)) as e:
-        assert not e.manifest.pristine_yaml_content["spack"]["config"]["ccache"]
+        assert not e.manifest.yaml_content["spack"]["config"]["ccache"]