This just tries to remove a few `deepcopy` calls. 

### Double deepcopy in `_write_section`

This call:
https://github.com/spack/spack/blob/54ee7d4165824a0e2fd81c91411f71ad50a8106a/lib/spack/spack/config.py#L178-L179

is redundant, since the first thing done within the `validate` function is to take a deepcopy of the input data. 

### Deepcopy when processing `definitions` in YAML files

This other call:
https://github.com/spack/spack/blob/54ee7d4165824a0e2fd81c91411f71ad50a8106a/lib/spack/spack/environment/environment.py#L960-L964

seems to be needed to `pop` from the input. The function is reworked to avoid the need for popping.

### Deepcopy on YAML validation

Before https://github.com/spack/spack/pull/20526 the validation of a YAML file was modifying the data (!), so we were using a defensive deepcopy to be able to round-trip data:

https://github.com/spack/spack/blob/54ee7d4165824a0e2fd81c91411f71ad50a8106a/lib/spack/spack/config.py#L1085

We don't need that since #20526 so here I removed the `deepcopy` and simplified `EnvironmentManifestFile`, which was still keeping track of a "pristine" YAML file and a YAML file with defaults attached.


---

Measured on the same benchmark as in https://github.com/spack/spack/pull/43745 on:

* **Spack:** 0.23.0.dev0 (15efcbe042692ebd6a634add316fb6564e26642c)
* **Python:** 3.11.5
* **Platform:** linux-ubuntu20.04-icelake
* **Concretizer:** clingo

```console
$ spack env create -d .
$ spack -e . add $(spack list)
$ time spack -e . find
% THE PR
==> 0 installed packages

real	0m3,125s
user	0m3,011s
sys	0m0,070s

% develop
==> 0 installed packages

real	0m3,934s
user	0m3,815s
sys	0m0,073s
```